разделил события и вынес в отдельные файлы
