+++ .gitignore
@@ -116,3 +116,4 @@ dmypy.json
 #Config
 .idea/
 test.py
+.vscode
