+++ events/__init__.py
@@ -1,0 +1,2 @@
+from .methods import *
+from .auth import *
